Week 1/bank/bank.py
# bank.py

def determine_greeting_cost(user_greeting):
    """
    Determine the cost of a user's greeting.

    Args:
        user_greeting (str): The user's greeting.

    Returns:
        int: The cost of the greeting based on specific conditions.
    """
    # TO DO: Implement the logic for determining the cost
    pass

if __name__ == "__main__":
    user_greeting = input("Enter a greeting: ")
    cost = determine_greeting_cost(user_greeting)
    print(f"The cost of your greeting is ${cost}.")
